Added capability to resize the window
And it seems to be broken under many conditions, but it compiles without
errors which feels like the important thing. If the rendering after
resize remains broken when we get actual shaders then I'll figure out
what the hell is going on there.

There is room for a lot of little improvements to be made here, but I'm
happy enough with how it works to push this.