Use vertex buffer instead of hardcoded shaders
The subject line isn't nearly enough info. This adds all of the memory
allocation and various supporting machinery to use vertex buffers
instead of the hardcoded shader values. To accomplish this it brings in
a whole new dependency with Vulkan Memory Allocator, and makes a mess of
a lot of the code base.

It compiles and runs without errors though, so we plow ahead.