Add index buffer to render and fix several bugs
We now use an index buffer for rendering, how sophisticated. More
importantly this fixes a couple bugs related to using pointers to
compound literals that have gone out of scope, which was causing
intermittent segfaults.

It also fixes the bug causing the image to not resize with the changes
in window size. That bug was caused by two things, not refreshing the
physical device capabilities on resize, and not rebuilding the graphics
pipeline and render pass when the dimensions of the window changed.